Fix several JSON-related bugs
Now `TRAILING_COMMA` reports on trailing comma in JSON.

A new error `TOO_MUCH_JSON` is added, which reports on useless
characters after end of a JSON expression.
e.g. `yee` in `{"x": 1}yee`

In JSON, word starts with `true` or `false` won't have these words
highlighted.
e.g. `truexxx`
In the past, `xxx` is marked as error and `true` is still marked as
boolean. Now the whole word is marked as error.